class-33/demo/things-api/things/serializer.py
from rest_framework import serializers
from .models import Things


class ThingsSerializer(serializers.ModelSerializer):
    class Meta:
        fields = ('id', 'owner', 'name', 'description', 'created_at', 'updated_at')
        model = Things
